refactor(validation): Separate file for validation logic
